requirement.txt
torch
torchvision
scikit-learn
matplotlib
Pillow
kagglehub
